+++ pkg/yappgen/assets/assets.go
@@ -1,0 +1,12 @@
+package assets
+
+import _ "embed"
+
+//go:embed YAPPgenerator_v3.scad
+var YAPPGeneratorSCAD []byte
+
+// GetGeneratorSCAD returns the embedded YAPP generator SCAD content.
+func GetGeneratorSCAD() []byte {
+	return YAPPGeneratorSCAD
+}
+